+++ README.md
@@ -1,3 +1,45 @@
-## ねこゲームを立ち上げる方法
-cd .\cat-game\
-python -m http.server 8000
+# ねこゲーム（React + Phaser）
+
+## 概要
+AIで動く「ねこ」と、プレイヤーが操作する「おもちゃ」で遊ぶゲームです。ねこはAIモデルで自律的に動き、プレイヤーはおもちゃを操作してねこと遊びます。
+
+## セットアップ手順
+
+1. Node.js（v18以上推奨）をインストールしてください。
+2. このディレクトリで依存パッケージをインストールします。
+   ```sh
+   npm install
+   ```
+3. 開発サーバーを起動します。
+   ```sh
+  $env:VERSION="v9"; npm run dev
+   ```
+4. ブラウザで `http://localhost:5173` を開くとゲームが遊べます。
+
+## ディレクトリ構成
+
+- `public/` : 画像・モデル・configファイルなど静的アセット
+- `src/`    : Reactコンポーネント・ゲームロジック
+
+## ゲームの流れ
+1. スタート画面で「スタート」ボタンを押すとゲーム画面に切り替わります。
+2. ゲーム画面では、AIねこが自律的に動きます。
+3. プレイヤーは矢印キーまたは画面下のボタンでおもちゃを操作します。
+4. 速度切替・モード切替ボタンで難易度や速さを調整できます。
+
+## 操作説明
+- 矢印キーまたは画面下のボタンでおもちゃを上下左右に動かせます。
+- 速度切替ボタンでおもちゃの速さを変更できます。
+- モード切替ボタンで難易度（ハード/イージー）を変更できます。
+
+## canvasとは？
+- canvas（キャンバス）は、HTMLで図形や画像、アニメーション、ゲーム画面などを描画するための領域です。
+- このゲームでは、Phaserというゲームエンジンがcanvas上にゲーム画面を描画しています。
+- Reactアプリの中にcanvasを埋め込むことで、UIとゲーム画面を組み合わせています。
+
+## デプロイ方法
+VercelやNetlifyなど静的サイトホスティングサービスで`$env:VERSION="v9"; npm run build`後の`dist/`フォルダをアップロードしてください。
+
+---
+
+何か問題があればIssueまたはDiscussionsでご連絡ください。
