+++ cat-game/public/cat.png
