+++ cat-game/public/style.css
@@ -1,0 +1,13 @@
+body {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  min-height: 100vh;
+  background-color: #f0f0f0;
+}
+
+#game-container {
+  position: relative;
+  border: 1px solid #ccc;
+  overflow: hidden;
+}
