+++ cat-game/public/toy.png
